session-controller: make take_device method public
We ultimately want to move the renderering related code
off to a separate file. When moved that code will need
to call back to the session controller to do that actual
talking to logind.

This commit exports the take_device method from the
session controller code, so that, in a later commit, the
rendering code can make use of it.