native: Add getter for seat id
We ultimately want to move the rendering related code
off to a separate file. When moved that code will need
the seat id, determined by the session controller, so
that it can identify the most appropriate GPU/card node
to use.

This commit adds a getter for the seat id.