native-renderer: add way to get modesetting fd
We get the kms fd ourselves, and then tell cogl about it.
Later whenever we need the kms fd we then ask cogl what it
is.  It would be good to get rid of cogl as a middle man.

This commit lays the ground work toward that end by adding
api to access kms fd set up by the native renderer.