src/Makefile.am: Fix building of stable wayland protcols
Currently there's not a single stable wayland protocol being build in mutter.
But when I recently tried to add the viewporter protocol, which is stable, it turned out to not work due to the assumption, that there's always a version number in protocls. Thats not true for stable protocols.

This fixes the issue and as a side effect makes the build process work for protocols with multi-word names, too, making it unnessecary to have extra cases for them.